fix and add fun delete

+++ src/main/java/home/local/vtbtest/controller/FileController.java
@@ -63,7 +63,7 @@ public class FileController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteFile(@PathVariable Integer id) {
+    public String deleteFile(@PathVariable Long id) {
         storageService.delete(id);
         return "redirect:/files/list";
     }
